+++ mayan/apps/permissions/methods.py
@@ -1,0 +1,23 @@
+from __future__ import unicode_literals
+
+from django.db import transaction
+
+from mayan.apps.user_management.events import event_group_edited
+
+
+def method_group_roles_add(self, queryset, _user):
+    with transaction.atomic():
+        event_group_edited.commit(
+            actor=_user, target=self
+        )
+        for role in queryset:
+            self.roles.add(role)
+
+
+def method_group_roles_remove(self, queryset, _user):
+    with transaction.atomic():
+        event_group_edited.commit(
+            actor=_user, target=self
+        )
+        for role in queryset:
+            self.roles.remove(role)